Refactor old export async methods to ActiveJobs [SCI-9092]
