+++ app/controllers/my_module_repositories_controller.rb
@@ -152,7 +152,14 @@ class MyModuleRepositoriesController < ApplicationController
 
   def export_repository
     if params[:header_ids]
-      RepositoryZipExport.generate_zip(params, @repository, current_user)
+      RepositoryZipExportJob.perform_later(
+        user_id: current_user.id,
+        params: {
+          repository_id: @repository.id,
+          my_module_id: @my_module.id,
+          header_ids: params[:header_ids]
+        }
+      )
 
       Activities::CreateActivityService.call(
         activity_type: :export_inventory_items_assigned_to_task,
