+++ app/controllers/repositories_controller.rb
@@ -337,7 +337,14 @@ class RepositoriesController < ApplicationController
 
   def export_repository
     if params[:row_ids] && params[:header_ids]
-      RepositoryZipExport.generate_zip(params, @repository, current_user)
+      RepositoryZipExportJob.perform_later(
+        user_id: current_user.id,
+        params: {
+          repository_id: @repository.id,
+          row_ids: params[:row_ids],
+          header_ids: params[:header_ids]
+        }
+      )
       log_activity(:export_inventory_items)
       render json: { message: t('zip_export.export_request_success') }
     else
@@ -359,7 +366,12 @@ class RepositoriesController < ApplicationController
   def export_repository_stock_items
     row_ids = @repository.repository_rows.where(id: params[:row_ids]).pluck(:id)
     if row_ids.any?
-      RepositoryStockLedgerZipExport.generate_zip(row_ids, current_user.id)
+      RepositoryStockZipExportJob.perform_later(
+        user_id: current_user.id,
+        params: {
+          repository_row_ids: row_ids
+        }
+      )
       render json: { message: t('zip_export.export_request_success') }
     else
       render json: { message: t('zip_export.export_error') }, status: :unprocessable_entity
