+++ app/services/repository_zip_export.rb
@@ -3,39 +3,6 @@
 require 'csv'
 
 module RepositoryZipExport
-  def self.generate_zip(params, repository, current_user)
-    # Fetch rows in the same order as in the currently viewed datatable
-    if params[:my_module_id]
-      rows = if repository.is_a?(RepositorySnapshot)
-               repository.repository_rows
-             else
-               repository.repository_rows
-                         .joins(:my_module_repository_rows)
-                         .where(my_module_repository_rows: { my_module_id: params[:my_module_id] })
-             end
-      if repository.has_stock_management?
-        rows = rows.left_joins(my_module_repository_rows: :repository_stock_unit_item)
-                   .select(
-                     'repository_rows.*',
-                     'my_module_repository_rows.stock_consumption'
-                   )
-      end
-    else
-      ordered_row_ids = params[:row_ids]
-      id_row_map = RepositoryRow.where(id: ordered_row_ids,
-                                       repository: repository)
-                                .index_by(&:id)
-      rows = ordered_row_ids.collect { |id| id_row_map[id.to_i] }
-    end
-
-    zip = ZipExport.create(user: current_user)
-    zip.generate_exportable_zip(
-      current_user.id,
-      to_csv(rows, params[:header_ids], current_user, repository, nil, params[:my_module_id].present?),
-      :repositories
-    )
-  end
-
   def self.to_csv(rows, column_ids, user, repository, handle_file_name_func = nil, in_module = false)
     # Parse column names
     csv_header = []
